+++ src/Contracts/Changelog.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ##### 2023-XX-YY
 
+- Added Laravel 10 support
 - Added the `DetailedAmount` interface
 
 ## 3.5.0